move to session_privilege_mgr
